ref: Reorginize common folder structure to have separate components and styled components directories
