+++ src/modules/common/components/CloseButton/CloseButton.styled.ts
