+++ src/modules/common/components/Contacts/ContactsForm.tsx
@@ -1,9 +1,12 @@
 import { Field, Form, Formik } from 'formik';
 import React from 'react';
-import { contactsValidationSchema } from '../../../schemas/contacts.schema';
-import contactsService from '../../../services/contacts.service';
-import { IContacts } from '../../../types/contacts';
-import { Button as StyledButton, ButtonWhite as StyledButtonWhite } from '../styled/button.styled';
+import { contactsValidationSchema } from '../../../../schemas/contacts.schema';
+import contactsService from '../../../../services/contacts.service';
+import { IContacts } from '../../../../types/contacts';
+import {
+  Button as StyledButton,
+  ButtonWhite as StyledButtonWhite
+} from '../../styled/button.styled';
 import * as Styled from './ContactsForm.styled';
 
 interface IProps {
