+++ src/modules/common/components/slider/Slider.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import MainSlideOne360 from '../../../assets/images/image-360-main-slide-1.png';
-import MainSlideOne from '../../../assets/images/image-main-slide-1.png';
+import MainSlideOne360 from '../../../../assets/images/image-360-main-slide-1.png';
+import MainSlideOne from '../../../../assets/images/image-main-slide-1.png';
 import ButtonGreen from '../ButtonGreen/ButtonGreen';
 
 const Slider: React.FC = () => (
