+++ src/modules/layout/Layout.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
 import Header from './header/Header';
-import BottomContacts from '../common/bottom-contacts/BottomContacts';
+import BottomContacts from '../common/components/bottom-contacts/BottomContacts';
 import Footer from './footer/Footer';
 
 const Layout: React.FC = () => (
